Sync scripts with the submitted paper
Brings the reproducibility package up to date with the manuscript as
submitted, and adds the four scripts the Supplementary Data depends on.

New:
  helper_supplement_data.py  builds Supplementary Data S1-S7
  kd_10_prediction_excel.py  kd_predictions.xlsx      -> S6
  kd_04d_grid_figures.py     kd_model_selection.xlsx  -> S5
  motif_26_cddm_pspa_kld.py  PSPA-reference KLD       -> S4

Updated:
  helper_supplement_tables.py  now the standalone dataset export to out/,
                               leaving supplement_table/ to the S1-S7 builder
  pathway_08_figure.py         median tick + per-method win rate on Fig. 8i-k
  plus the motif / scoring / kd / pathway scripts already carried in the
  working tree, now matching the source repo exactly

README: script count, a per-workbook S1-S7 table with upstream producers,
the new scripts in the run order, three artifact-manifest rows, and
supplement_table/ in the layout. .gitignore ignores the generated workbooks.

Verified against the source repo by re-running the pipeline here with this
repo's own trimmed packages: motif_16, motif_24, motif_25 and the KLD script
reproduce byte-identical outputs, and all seven workbooks come out
value-identical sheet by sheet.

+++ nbs/data_01_preprocess_dataset.py
@@ -201,7 +201,7 @@ def load_elm():
 
     elm = map_substrate(RAW / 'idmapping_2025_03_12_elm.xlsx', elm, 'substrate_uniprot')
 
-    elm['site'] = elm['acceptor'] + elm['position'].astype(str)
+    elm['site'] = elm['acceptor'] + elm['position'].astype(int).astype(str)   # int cast for parity with load_gps (guards float 'S140.0')
     elm = filter_valid_sites(elm)
 
     elm = elm[['kinase', 'kinase_uniprot', 'kinase_genes', 'substrate_uniprot',
@@ -417,6 +417,7 @@ def build_ks_dataset(df_grouped):
     for out_col, info_col in KINASE_INFO_MAP.items():
         df[out_col] = clean.map(info[info_col])
 
+    # NB: counts raw kinase_uniprot, so the 3 isoform accessions (PRKCB P05771-2, LYN P07948-2, PRKG1 Q13976-2) count separately from their canonical form - a small known inconsistency (~1,071 sites inflated, ~11 crossing the thr=40 CDDM cutoff), left as-is to avoid a full-pipeline rebuild.
     df['num_kin'] = df.groupby('sub_site')['kinase_uniprot'].transform('nunique')
 
     print('  kinases on the kinome tree:', int(df.kinase_on_tree.sum()), '/', len(df), 'pairs')
